refactor(auth): make (authenticated) layout passthrough and wrap all pages with AuthGate
Phase 2: Next.js 16 の cacheComponents 有効化準備として、認証層の責務を
layout から各 page にある AuthGate に移譲する。

主な変更:
- (authenticated)/layout.tsx を passthrough（children を返すだけ）化。
  これにより layout が cookies を読まなくなり、cacheComponents 有効化時に
  static shell が dynamic source 由来で潰れるのを防ぐ
- (authenticated) 配下の全 page を `<AuthGate redirectTo={...}>` でラップ
- 既に getCurrentUser() で initialUser を構築している mypage / profile/edit は
  既存ロジックを維持（getCachedUserInfo の cache HIT で実質コストゼロ）
- subscription/checkout は Client Component だったため、Checkout.tsx に分離して
  page.tsx を Server Component ラッパーに変更し AuthGate でラップ

cacheComponents は未有効。本コミット時点では PPR 効果は出ないが、
全 (authenticated) page が cacheComponents 互換になっている。
