+++ frontend/src/app/[locale]/(authenticated)/logout/page.tsx
@@ -1,5 +1,15 @@
+import { AuthGate } from "@/components/shared/auth";
 import { Logout } from "./Logout";
 
-export default async function LogoutPage() {
-  return <Logout />;
+export default async function LogoutPage({
+  params,
+}: {
+  params: Promise<{ locale: string }>;
+}) {
+  const { locale } = await params;
+  return (
+    <AuthGate redirectTo={`/${locale}/login`}>
+      <Logout />
+    </AuthGate>
+  );
 }
