+++ frontend/src/app/[locale]/(authenticated)/mypage/page.tsx
@@ -36,7 +36,6 @@ export default async function Page({
 
   const user = await getCurrentUser();
 
-  // 認証チェック（Layoutでも行っているが、ビルド時や型解決のために必要）
   if (!user) {
     redirect(loginPath);
   }
