+++ frontend/src/app/[locale]/(authenticated)/personal/calendar/page.tsx
@@ -1,5 +1,6 @@
 import type { Metadata } from "next";
 import { getTranslations } from "next-intl/server";
+import { AuthGate } from "@/components/shared/auth";
 import { MinimalLayout } from "@/components/shared/layouts/MinimalLayout";
 import { buildMetadata } from "@/lib/metadata";
 import { PersonalCalendar } from "./PersonalCalendar";
@@ -27,11 +28,13 @@ export default async function Page({
   const t = await getTranslations({ locale, namespace: "personalCalendar" });
 
   return (
-    <MinimalLayout
-      backHref={`/${locale}/personal/pages`}
-      headerTitle={t("title")}
-    >
-      <PersonalCalendar />
-    </MinimalLayout>
+    <AuthGate redirectTo={`/${locale}/login`}>
+      <MinimalLayout
+        backHref={`/${locale}/personal/pages`}
+        headerTitle={t("title")}
+      >
+        <PersonalCalendar />
+      </MinimalLayout>
+    </AuthGate>
   );
 }
