+++ frontend/src/app/[locale]/(authenticated)/personal/pages/[page_id]/edit/page.tsx
@@ -1,5 +1,6 @@
 import type { Metadata } from "next";
 import { getTranslations } from "next-intl/server";
+import { AuthGate } from "@/components/shared/auth";
 import { buildMetadata } from "@/lib/metadata";
 import { PageEdit } from "./PageEdit";
 
@@ -16,6 +17,15 @@ export async function generateMetadata({
   });
 }
 
-export default async function PersonalPageEditPage() {
-  return <PageEdit />;
+export default async function PersonalPageEditPage({
+  params,
+}: {
+  params: Promise<{ locale: string; page_id: string }>;
+}) {
+  const { locale } = await params;
+  return (
+    <AuthGate redirectTo={`/${locale}/login`}>
+      <PageEdit />
+    </AuthGate>
+  );
 }
