+++ frontend/src/app/[locale]/(authenticated)/personal/pages/[page_id]/page.tsx
@@ -1,3 +1,4 @@
+import { AuthGate } from "@/components/shared/auth";
 import { DefaultLayout } from "@/components/shared/layouts/DefaultLayout";
 import { buildMetadata } from "@/lib/metadata";
 import { PageDetail } from "./PageDetail";
@@ -7,10 +8,17 @@ export const metadata = buildMetadata({
   description: "稽古ページの詳細を確認できます。",
 });
 
-export default async function Page() {
+export default async function Page({
+  params,
+}: {
+  params: Promise<{ locale: string; page_id: string }>;
+}) {
+  const { locale } = await params;
   return (
-    <DefaultLayout>
-      <PageDetail />
-    </DefaultLayout>
+    <AuthGate redirectTo={`/${locale}/login`}>
+      <DefaultLayout>
+        <PageDetail />
+      </DefaultLayout>
+    </AuthGate>
   );
 }
