+++ frontend/src/app/[locale]/(authenticated)/settings/email/page.tsx
@@ -1,3 +1,4 @@
+import { AuthGate } from "@/components/shared/auth";
 import { EmailSetting } from "./EmailSetting";
 
 export default async function Page({
@@ -7,5 +8,9 @@ export default async function Page({
 }) {
   const { locale } = await params;
 
-  return <EmailSetting locale={locale} />;
+  return (
+    <AuthGate redirectTo={`/${locale}/login`}>
+      <EmailSetting locale={locale} />
+    </AuthGate>
+  );
 }
