+++ frontend/src/app/[locale]/(authenticated)/settings/publicity/page.tsx
@@ -1,3 +1,4 @@
+import { AuthGate } from "@/components/shared/auth";
 import { PublicitySetting } from "./PublicitySetting";
 
 export default async function Page({
@@ -7,5 +8,9 @@ export default async function Page({
 }) {
   const { locale } = await params;
 
-  return <PublicitySetting locale={locale} />;
+  return (
+    <AuthGate redirectTo={`/${locale}/login`}>
+      <PublicitySetting locale={locale} />
+    </AuthGate>
+  );
 }
