+++ frontend/src/app/[locale]/(authenticated)/settings/tags/page.tsx
@@ -1,3 +1,4 @@
+import { AuthGate } from "@/components/shared/auth";
 import { TagManagement } from "./TagManagement";
 
 export default async function Page({
@@ -7,5 +8,9 @@ export default async function Page({
 }) {
   const { locale } = await params;
 
-  return <TagManagement locale={locale} />;
+  return (
+    <AuthGate redirectTo={`/${locale}/login`}>
+      <TagManagement locale={locale} />
+    </AuthGate>
+  );
 }
