+++ frontend/src/app/[locale]/(authenticated)/social/posts/[post_id]/edit/page.tsx
@@ -1,5 +1,6 @@
 import type { Metadata } from "next";
 import { getTranslations } from "next-intl/server";
+import { AuthGate } from "@/components/shared/auth";
 import { buildMetadata } from "@/lib/metadata";
 import { SocialPostEdit } from "./SocialPostEdit";
 
@@ -16,6 +17,15 @@ export async function generateMetadata({
   });
 }
 
-export default async function SocialPostEditPage() {
-  return <SocialPostEdit />;
+export default async function SocialPostEditPage({
+  params,
+}: {
+  params: Promise<{ locale: string; post_id: string }>;
+}) {
+  const { locale } = await params;
+  return (
+    <AuthGate redirectTo={`/${locale}/login`}>
+      <SocialPostEdit />
+    </AuthGate>
+  );
 }
