+++ frontend/src/app/[locale]/(authenticated)/social/posts/new/page.tsx
@@ -1,5 +1,6 @@
 import type { Metadata } from "next";
 import { getTranslations } from "next-intl/server";
+import { AuthGate } from "@/components/shared/auth";
 import { buildMetadata } from "@/lib/metadata";
 import { SocialPostCreate } from "./SocialPostCreate";
 
@@ -16,6 +17,15 @@ export async function generateMetadata({
   });
 }
 
-export default async function SocialPostNewPage() {
-  return <SocialPostCreate />;
+export default async function SocialPostNewPage({
+  params,
+}: {
+  params: Promise<{ locale: string }>;
+}) {
+  const { locale } = await params;
+  return (
+    <AuthGate redirectTo={`/${locale}/login`}>
+      <SocialPostCreate />
+    </AuthGate>
+  );
 }
