+++ frontend/src/app/[locale]/(authenticated)/social/posts/search/page.tsx
@@ -1,5 +1,6 @@
 import type { Metadata } from "next";
 import { getTranslations } from "next-intl/server";
+import { AuthGate } from "@/components/shared/auth";
 import { buildMetadata } from "@/lib/metadata";
 import { SocialSearch } from "./SocialSearch";
 
@@ -16,6 +17,15 @@ export async function generateMetadata({
   });
 }
 
-export default async function SocialSearchPage() {
-  return <SocialSearch />;
+export default async function SocialSearchPage({
+  params,
+}: {
+  params: Promise<{ locale: string }>;
+}) {
+  const { locale } = await params;
+  return (
+    <AuthGate redirectTo={`/${locale}/login`}>
+      <SocialSearch />
+    </AuthGate>
+  );
 }
